Search now use Symfony\Contracts\EventDispatcher\EventDispatcherInterface if available
